+++ packages/react-storage/src/components/FileUploader/utils/__tests__/getInput.spec.ts
@@ -20,13 +20,11 @@ const processFile: GetInputParams['processFile'] = ({ key, ...rest }) => ({
 
 const stringPath = 'my-path/';
 
-const onProcessFileSuccess = jest.fn();
 const inputBase: Omit<GetInputParams, 'path' | 'accessLevel'> = {
   file,
   key,
   onProgress,
   processFile: undefined,
-  onProcessFileSuccess,
 };
 const pathStringInput: GetInputParams = {
   ...inputBase,
@@ -55,7 +53,6 @@ const accessLevelWithPathInput: GetInputParams = {
 
 describe('getInput', () => {
   beforeEach(() => {
-    onProcessFileSuccess.mockClear();
     fetchAuthSpy.mockClear();
   });
 
@@ -151,33 +148,6 @@ describe('getInput', () => {
     expect(output).toStrictEqual(expected);
   });
 
-  it('calls `onProcessFileSuccess` when `processFile` is provided', async () => {
-    const processedKey = `processedKey`;
-
-    const input = getInput({
-      ...pathStringInput,
-      processFile: ({ key: _, ...rest }) => ({
-        key: processedKey,
-        ...rest,
-      }),
-    });
-
-    await input();
-
-    expect(onProcessFileSuccess).toHaveBeenCalledTimes(1);
-    expect(onProcessFileSuccess).toHaveBeenCalledWith({
-      processedKey,
-    });
-  });
-
-  it('does not call `onProcessFileSuccess` when `processFile` is not provided', async () => {
-    const input = getInput(pathStringInput);
-
-    await input();
-
-    expect(onProcessFileSuccess).not.toHaveBeenCalled();
-  });
-
   it('includes additional values returned from `processFile` in `options`', async () => {
     const contentDisposition = 'attachment';
     const metadata = { key };
@@ -211,33 +181,6 @@ describe('getInput', () => {
     expect(output.options?.contentDisposition).toStrictEqual(
       contentDisposition
     );
-  });
-
-  it('calls `onProcessFileSuccess` after fetchAuthSession', async () => {
-    const processedKey = `processedKey`;
-
-    const input = getInput({
-      ...pathCallbackInput,
-      processFile: ({ key: _, ...rest }) => ({
-        key: processedKey,
-        ...rest,
-      }),
-    });
-
-    await input();
-
-    const fetchAuthSessionCallOrder = fetchAuthSpy.mock.invocationCallOrder[0];
-    const onProcessFileSuccessCallORder =
-      onProcessFileSuccess.mock.invocationCallOrder[0];
-    expect(fetchAuthSessionCallOrder).toBeLessThan(
-      onProcessFileSuccessCallORder
-    );
-
-    expect(fetchAuthSpy).toHaveBeenCalledTimes(1);
-    expect(onProcessFileSuccess).toHaveBeenCalledTimes(1);
-    expect(onProcessFileSuccess).toHaveBeenCalledWith({
-      processedKey,
-    });
   });
 
   it('defaults `options.contentType` to "binary/octet-stream" when no file type is provided', async () => {