+++ packages/react-storage/src/components/index.ts
@@ -1,8 +1,13 @@
-export { FileUploader, FileUploaderProps } from './FileUploader';
+export {
+  FileUploader,
+  FileUploaderHandle,
+  FileUploaderProps,
+} from './FileUploader';
 
 export { StorageImage, StorageImageProps } from './StorageImage';
 export {
   StorageManager,
+  StorageManagerHandle,
   StorageManagerProps,
   DropZoneProps,
   ContainerProps,