+++ packages/react-storage/src/index.ts
@@ -1,9 +1,11 @@
 export {
   FileUploader,
+  FileUploaderHandle,
   FileUploaderProps,
   StorageImage,
   StorageImageProps,
   StorageManager,
+  StorageManagerHandle,
   StorageManagerProps,
   DropZoneProps,
   ContainerProps,